add: NotificationModalWrapper.jsx; finish: WebsiteWithMobileAppCalculator.jsx
