+++ src/pages/services.jsx
@@ -1,13 +1,15 @@
-import React from 'react';
-import Layout from "@/components/Layout";
+import React, {useState} from 'react';
 import HeaderOrder from "@/components/Order/HeaderOrder";
 import Calculator from "@/components/Services/Calculator";
+import NotificationModalWrapper from "@/components/NotificationModalWrapper";
 const Services = () => {
+    const [isModalOpen, setIsModalOpen] = useState(false)
+
     return (
-        <Layout>
+        <NotificationModalWrapper isModalOpen={isModalOpen} setIsModalOpen={setIsModalOpen}>
             <HeaderOrder/>
-            <Calculator />
-        </Layout>
+            <Calculator setIsModalOpen={setIsModalOpen}/>
+        </NotificationModalWrapper>
     );
 };
